requirements.txt
django==5.2
djangorestframework==3.16
djangorestframework-simplejwt==5.3
python-decouple==3.8
django-cors-headers==4.3
django-filter==23.5
dj-database-url==2.2
gunicorn>=25.3.0
